Phase 1: split triggersrepo, extract DLQ, generic fanout
- triggersrepo.go (793 LOC, 23 methods) split into:
  repository.go (type + ctor + cipher + tx helpers),
  triggers.go (CRUD), subscriptions.go (CRUD),
  failures.go (circuit-breaker), secrets.go (rotation),
  internal.go (cache/dispatcher reads, no auth scoping).

- internal/nats/bridge.go split: DLQ + helper subject derivation moved
  to dlq.go. bridge.go now only handles the Kafka->NATS republish path,
  which is itself transitional and will be deleted at exclusive cutover.

- metriclistener: generic fanoutAndFire + evalAndFire in fire.go
  replaces the duplicated processSingle* / process*Webhook templates.
  signal.go and events.go now own only the per-type entry point and
  payload builder; the eval + permission-denied + fire flow is the
  same code path.

Pure refactor; behavior identical. All race tests pass.
